+++ src/cli/mod.rs
@@ -1,0 +1,6 @@
+pub mod progress;
+pub mod commands;
+
+pub use commands::*;
+#[allow(unused_imports)]
+pub use progress::{CliProgressBar, demo_progress_indicator};
